requirements.txt
requests==2.31.0
protobuf==4.23.2
gtfs-realtime-bindings==1.0.0
python-dotenv==1.0.0
pandas==2.0.3